change: restrict origins for CSRF protection
We want to prevent that malicious websites can do a cross site request forgery attack (CSRF) on our service. I.e. a malicious website trying to use authentication from a trusted website, but executing arbitrary api calls with the gained credentials.

Note: checking the content type in the middleware does not add any benefit, as a simple request as well as a preflight request from the wrong origin is already rejected by the CORS middleware